requirements.txt
# HTTP & scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.44.0
tenacity>=8.3.0
pyyaml>=6.0.1

# LLM clients
anthropic>=0.29.0
openai>=1.35.0

# Data & storage
sqlalchemy>=2.0.0
pyarrow>=16.0.0
datasets>=2.20.0

# Settings
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Filtering
langdetect>=1.0.9
